fix: complete OGC seam split type relocation (#1579)
The seam-split commit moved shared OGC GeoJSON/geometry types into the
Honua.Protocols.Ogc.Common namespace (Ogc.Shared project) but left several
OgcApi consumers, the GeoJSON benchmark, and several test fixtures importing
them from their former Ogc.Api.Features.Models/Services namespaces, so the
solution did not compile.

- Relocate OgcGeoJsonGeometryShapeValidator (and GeometryShapeValidationResult)
  into Ogc.Shared/Common so OgcFeaturesGeometryServices, which now lives in the
  shared assembly, can reach it without a dependency back into OgcApi.
- Add the Honua.Protocols.Ogc.Common using to the OgcApi dependency/registration
  and mutation-helper files and to the benchmark and affected test files.
- Point the OgcGeoJsonFeature test aliases at the new Ogc.Common namespace.
