+++ tests/dotnet/Honua.Protocols.OgcApi.Tests/Source/Features/OgcFeaturesSpatialReferenceTests.cs
@@ -8,6 +8,7 @@ using System.Text.Json;
 using FluentAssertions;
 using Honua.Protocols.Ogc.Api.Features;
 using Honua.Protocols.Ogc.Api.Features.Models;
+using Honua.Protocols.Ogc.Common;
 using Honua.TestKit;
 using Honua.TestKit.Attributes;
 using Honua.TestKit.Constants;
