+++ tests/dotnet/Honua.Server.Tests/Features/Security/ServiceRbacAuthorizationTests.cs
@@ -36,7 +36,7 @@ using MetadataV2ServiceProtocols = Honua.Core.Features.Metadata.Domain.V2.Servic
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using OgcGeoJsonFeature = Honua.Protocols.Ogc.Api.Features.Models.GeoJsonFeature;
+using OgcGeoJsonFeature = Honua.Protocols.Ogc.Common.GeoJsonFeature;
 
 namespace Honua.Server.Tests.Features.Security;
